Added instructions for resolving libadwaita issues
As of 8/12/2022, a build dependency update in libadwaita (gtk4 >= 4.9.1) results in the build failing on Fedora 37 and Ubuntu 22.04. I've added some instructions to checkout to an earlier commit, so that the build can proceed.

This should be a temporary change, for until the updated dependencies are available in the Ubuntu and Fedora stable repos. This is also a bodge; if there are any other better ways to handle this, please let me know.